refactor(bin-designer): split pathGeometry into focused modules
Decompose the 690-LOC pathGeometry.ts into a slim facade plus three
focused sibling modules. Public API preserved via re-exports; no
behavior change.

- pathGeometryBezier.ts (153 LOC): Point2D, cubicBezier, flattenPath
  (the De Casteljau subdivision math).
- pathGeometryTriangulation.ts (127 LOC): triangulatePath via earclip,
  signedArea, pointInTriangle.
- pathGeometrySegments.ts (261 LOC): findNearestSegment,
  evaluateSegmentPoint, flattenSegment, splitBezierSegment,
  isSelfIntersecting + their internal helpers.
- pathGeometry.ts (213 LOC): facade re-exports + simple point-array
  helpers (bounds, scale, translate, insert/remove/update, snap,
  corner, isNear, clampPathToBounds, enforceSymmetry).
